주문 내역 화면 ui layer 결합

+++ app/src/main/java/com/woowa/banchan/domain/usecase/order/GetOrderLineItemUseCase.kt
@@ -12,7 +12,7 @@ class GetOrderLineItemUseCase @Inject constructor(
     private val orderRepository: OrderRepository
 ) {
 
-    operator fun invoke(orderId: Int) = flow {
+    operator fun invoke(orderId: Long) = flow {
         orderRepository.getOrderLineItem(orderId).collect { emit(it) }
     }.catch {
         emit(Result.failure(NotFoundProductsException()))
